+++ graphics/src/Window.cpp
@@ -318,12 +318,12 @@ void Window::present( const Image& image )
         return;
 
     float w, h;
-    if ( !m_Texture || SDL_GetTextureSize( m_Texture, &w, &h ) && ( w != static_cast<float>( image.getWidth() ) || h != static_cast<float>( image.getHeight() ) ) )
+    if ( !m_Texture || ( SDL_GetTextureSize( m_Texture, &w, &h ) && ( w != static_cast<float>( image.getWidth() ) || h != static_cast<float>( image.getHeight() ) ) ) )
     {
         if ( m_Texture )
             SDL_DestroyTexture( m_Texture );
 
-        m_Texture = SDL_CreateTexture( m_Renderer, SDL_PIXELFORMAT_RGBA32, SDL_TEXTUREACCESS_STREAMING, static_cast<int>( image.getWidth() ), static_cast<int>( image.getHeight() ) );
+        m_Texture = SDL_CreateTexture( m_Renderer, SDL_PIXELFORMAT_RGBA32, SDL_TEXTUREACCESS_STREAMING, image.getWidth(), image.getHeight() );
         if ( !m_Texture )
         {
             SDL_LogError( SDL_LOG_CATEGORY_APPLICATION, "Failed to create texture: %s", SDL_GetError() );
